# stuplus.api/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Stuplus API",
    "description": "Stuplus API Documentation."
  },
  "servers": [
    {
      "url": "http://212.98.224.208:25010/",
      "description": "Canli API"
    },
    {
      "url": "http://127.0.0.1:25010/",
      "description": "Local API"
    }
  ],
...
  "paths": {
...
    "/search/people": {
      "post": {
        "tags": [
          "Search"
        ],
        "description": "Get searched results(only users).",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchPeopleResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SearchPeopleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchPeopleRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/SearchPeopleRequest"
              }
            }
          }
        }
      }
    },
    "/search/groupChats": {
      "post": {
        "tags": [
          "Search"
        ],
        "description": "Get searched results(only group chats).",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchGroupChatsResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SearchGroupChatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchGroupChatsRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/SearchGroupChatsRequest"
              }
            }
          }
        }
      }
    },
    "/search/peopleAndGroupChats": {
      "post": {
        "tags": [
          "Search"
        ],
        "description": "Get searched results(people and group chats).",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchPeopleAndGroupChatsResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SearchPeopleAndGroupChatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchPeopleAndGroupChatsRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/SearchPeopleAndGroupChatsRequest"
              }
            }
          }
        }
      }
    },
    "/search/hashtag": {
      "post": {
        "tags": [
          "Search"
        ],
        "description": "Get hashtags by search term.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchHashtagResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SearchHashtagResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchHashtagRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/SearchHashtagRequest"
              }
            }
          }
        }
      }
    },
    "/search/question": {
      "post": {
        "tags": [
          "Search"
        ],
        "description": "Get searched questions.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchQuestionResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SearchQuestionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQuestionRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/SearchQuestionRequest"
              }
            }
          }
        }
      }
    },
    "/search/announcement": {
      "post": {
        "tags": [
          "Search"
        ],
        "description": "Get searched questions.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchQuestionResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SearchQuestionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQuestionRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/SearchQuestionRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SearchPeopleRequest": {
        "type": "object",
        "properties": {
          "searchTerm": {
            "type": "string",
            "example": "test"
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          }
        },
        "required": [
          "searchTerm",
          "page",
          "pageSize"
        ],
        "xml": {
          "name": "SearchPeopleRequest"
        }
      },
      "SearchPeopleResponse": {
        "type": "object",
        "properties": {
          "hasError": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "boolean"
              },
              "description": {
                "type": "string",
                "example": "Indicates whether the operation was successful or not."
              }
            }
          },
          "validationErrors": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "array"
              },
              "description": {
                "type": "string",
                "example": "List of validation errors."
              }
            }
          },
          "message": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              },
              "description": {
                "type": "string",
                "example": "Message describing the error."
              }
            }
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "62ab41a460884a883ed78080"
                },
                "username": {
                  "type": "string",
                  "example": "test1013"
                },
                "firstName": {
                  "type": "string",
                  "example": "test1013"
                },
                "lastName": {
                  "type": "string",
                  "example": "test1013"
                },
                "profilePhotoUrl": {},
                "__t": {
                  "type": "string",
                  "example": "User"
                },
                "confidenceScore": {
                  "type": "number",
                  "example": 42.900000000000006
                }
              }
            }
          }
        },
        "xml": {
          "name": "SearchPeopleResponse"
        }
      },
      "SearchGroupChatsRequest": {
        "type": "object",
        "properties": {
          "searchTerm": {
            "type": "string",
            "example": "uni"
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          }
        },
        "required": [
          "searchTerm",
          "page",
          "pageSize"
        ],
        "xml": {
          "name": "SearchGroupChatsRequest"
        }
      },
      "SearchGroupChatsResponse": {
        "type": "object",
        "properties": {
          "hasError": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "boolean"
              },
              "description": {
                "type": "string",
                "example": "Indicates whether the operation was successful or not."
              }
            }
          },
          "validationErrors": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "array"
              },
              "description": {
                "type": "string",
                "example": "List of validation errors."
              }
            }
          },
          "message": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              },
              "description": {
                "type": "string",
                "example": "Message describing the error."
              }
            }
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "62cc9ca25c9816cc1e36880d"
                },
                "title": {
                  "type": "string",
                  "example": "Muhasebe ve Vergi Uygulamaları 2. Sınıf"
                },
                "type": {
                  "type": "number",
                  "example": 0
                },
                "coverImageUrl": {
                  "type": "string",
                  "example": "https://stuplus-bucket.s3.eu-central-1.amazonaws.com/public/department_images/1657574858083-4379-muhabasebe.jpg"
                },
                "about": {
                  "type": "string",
                  "example": "Bu grubu sizin için biz(stuplus) oluşturduk. Tepe tepe kullanıp hayrını görün."
                },
                "hashTags": {
                  "type": "array",
                  "example": [
                    "muhasebevevergiuygulamalari2sinif",
                    "Muhasebe",
                    "ve",
                    "Vergi",
                    "Uygulamaları",
                    "2.",
                    "Sınıf",
                    "ataturkuniversitesi",
                    "Atatürk",
                    "Üniversitesi",
                    ""
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "schoolId": {
                  "type": "string",
                  "example": "62cc9c8f5c9816cc1e36858d"
                },
                "departmentId": {
                  "type": "string",
                  "example": "62cc9ca25c9816cc1e3687f9"
                },
                "grade": {
                  "type": "number",
                  "example": 2
                },
                "__t": {
                  "type": "string",
                  "example": "GroupChat"
                },
                "confidenceScore": {
                  "type": "number",
                  "example": 3.3000000000000003
                },
                "school": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "62cc9c8f5c9816cc1e36858d"
                    },
                    "emailFormat": {
                      "type": "string",
                      "example": "atauni.edu.tr"
                    },
                    "title": {
                      "type": "string",
                      "example": "Atatürk Üniversitesi"
                    },
                    "coverImageUrl": {
                      "type": "string",
                      "example": "https://stuplus-bucket.s3.eu-central-1.amazonaws.com/public/school_images/1657578697785-2773-ataturk_uni.png"
                    },
                    "type": {
                      "type": "number",
                      "example": 0
                    },
                    "recordStatus": {
                      "type": "number",
                      "example": 1
                    },
                    "__t": {
                      "type": "string",
                      "example": "School"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2022-07-11T22:35:13.961Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2022-07-11T22:35:13.961Z"
                    }
                  }
                },
                "department": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "62cc9ca25c9816cc1e3687f9"
                    },
                    "facultyId": {},
                    "schoolId": {
                      "type": "string",
                      "example": "62cc9c8f5c9816cc1e36858d"
                    },
                    "grade": {
                      "type": "number",
                      "example": 2
                    },
                    "title": {
                      "type": "string",
                      "example": "Muhasebe ve Vergi Uygulamaları"
                    },
                    "coverImageUrl": {
                      "type": "string",
                      "example": "https://stuplus-bucket.s3.eu-central-1.amazonaws.com/public/department_images/1657574858083-4379-muhabasebe.jpg"
                    },
                    "about": {
                      "type": "string",
                      "example": "Bu grubu sizin için biz(stuplus) oluşturduk. Tepe tepe kullanıp hayrını görün."
                    },
                    "avatarKey": {},
                    "recordStatus": {
                      "type": "number",
                      "example": 1
                    },
                    "__t": {
                      "type": "string",
                      "example": "Department"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2022-07-11T21:56:50.330Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2022-07-11T21:56:50.330Z"
                    }
                  }
                }
              }
            }
          }
        },
        "xml": {
          "name": "SearchGroupChatsResponse"
        }
      },
...
      "SearchPeopleAndGroupChatsResponse": {
        "type": "object",
        "properties": {
          "hasError": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "boolean"
              },
              "description": {
                "type": "string",
                "example": "Indicates whether the operation was successful or not."
              }
            }
          },
          "validationErrors": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "array"
              },
              "description": {
                "type": "string",
                "example": "List of validation errors."
              }
            }
          },
          "message": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              },
              "description": {
                "type": "string",
                "example": "Message describing the error."
              }
            }
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "62cc9c945c9816cc1e36860e"
                },
                "title": {
                  "type": "string",
                  "example": "Kamu Yönetimi 2. Sınıf"
                },
                "type": {
                  "type": "number",
                  "example": 0
                },
                "coverImageUrl": {
                  "type": "string",
                  "example": "https://stuplus-bucket.s3.eu-central-1.amazonaws.com/public/department_images/1657570576441-4039-kamu%20y%C3%83%C2%B6netimi.jpg"
                },
                "about": {
                  "type": "string",
                  "example": "Bu grubu sizin için biz(stuplus) oluşturduk. Tepe tepe kullanıp hayrını görün."
                },
                "hashTags": {
                  "type": "array",
                  "example": [
                    "kamuyonetimi2sinif",
                    "Kamu",
                    "Yönetimi",
                    "2.",
                    "Sınıf",
                    "ataturkuniversitesi",
                    "Atatürk",
                    "Üniversitesi",
                    ""
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "schoolId": {
                  "type": "string",
                  "example": "62cc9c8f5c9816cc1e36858d"
                },
                "departmentId": {
                  "type": "string",
                  "example": "62cc9c945c9816cc1e368604"
                },
                "grade": {
                  "type": "number",
                  "example": 2
                },
                "__t": {
                  "type": "string",
                  "example": "GroupChat"
                },
                "confidenceScore": {
                  "type": "number",
                  "example": 3.3000000000000003
                },
                "school": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "62cc9c8f5c9816cc1e36858d"
                    },
                    "emailFormat": {
                      "type": "string",
                      "example": "atauni.edu.tr"
                    },
                    "title": {
                      "type": "string",
                      "example": "Atatürk Üniversitesi"
                    },
                    "coverImageUrl": {
                      "type": "string",
                      "example": "https://stuplus-bucket.s3.eu-central-1.amazonaws.com/public/school_images/1657578697785-2773-ataturk_uni.png"
                    },
                    "type": {
                      "type": "number",
                      "example": 0
                    },
                    "recordStatus": {
                      "type": "number",
                      "example": 1
                    },
                    "__t": {
                      "type": "string",
                      "example": "School"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2022-07-11T22:35:13.961Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2022-07-11T22:35:13.961Z"
                    }
                  }
                },
                "department": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "62cc9c945c9816cc1e368604"
                    },
                    "facultyId": {},
                    "schoolId": {
                      "type": "string",
                      "example": "62cc9c8f5c9816cc1e36858d"
                    },
                    "grade": {
                      "type": "number",
                      "example": 4
                    },
                    "title": {
                      "type": "string",
                      "example": "Kamu Yönetimi"
                    },
                    "coverImageUrl": {
                      "type": "string",
                      "example": "https://stuplus-bucket.s3.eu-central-1.amazonaws.com/public/department_images/1657570576441-4039-kamu%20y%C3%83%C2%B6netimi.jpg"
                    },
                    "about": {
                      "type": "string",
                      "example": "Bu grubu sizin için biz(stuplus) oluşturduk. Tepe tepe kullanıp hayrını görün."
                    },
                    "avatarKey": {},
                    "recordStatus": {
                      "type": "number",
                      "example": 1
                    },
                    "__t": {
                      "type": "string",
                      "example": "Department"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2022-07-11T21:56:36.162Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2022-07-11T21:56:36.162Z"
                    }
                  }
                }
              }
            }
          }
        },
        "xml": {
          "name": "SearchPeopleAndGroupChatsResponse"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}